修改：【user】 变更用户信息相关处理

+++ Suyaa.Apis/Entities/ICreationEntity.cs
@@ -18,7 +18,7 @@ namespace Suyaa.Apis.Entities
         /// <summary>
         /// 创建用户编号
         /// </summary>
-        int CreatorUserNo { get; set; }
+        long CreatorUserNo { get; set; }
 
         /// <summary>
         /// 创建时间
