+++ Suyaa.Apis/Entities/IDeletionEntity.cs
@@ -18,7 +18,7 @@ namespace Suyaa.Apis.Entities
         /// <summary>
         /// 删除用户编号
         /// </summary>
-        int? DeleterUserNo { get; set; }
+        long? DeleterUserNo { get; set; }
 
         /// <summary>
         /// 删除时间
